websocket_server/requirements.txt
autobahn==0.16.0
pygame==1.9.2b8
PyYAML==3.11
